requirements.txt
Django==3.1
django-cors-headers==3.4.0
django-decouple==2.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5